refactored unaryCall to now be generateUnaryCall which is a factory for call that extends unary call proto
