+++ lib/compose.js
@@ -1,4 +1,4 @@
-const generateUnaryCall = require("./calls/unaryCall");
+const generateUnaryCall = require("./calls/generateUnaryCall");
 const generateClientStreamCall = require("./calls/generateClientStreamCall");
 const ServerStreamCall = require("./calls/serverStreamCall");
 const generateDuplex = require("./calls/generateDuplex");
